+++ async-stripe/src/pagination.rs
@@ -1,84 +1,110 @@
 // Necessary under tokio-blocking since `Response` is a type alias to a `Result`
 #![allow(clippy::missing_errors_doc)]
-use serde::de::DeserializeOwned;
 use serde::Serialize;
-use stripe_types::{AsCursorOpt, List, Object};
+use stripe_types::{List, PaginableList, SearchList};
 
-use crate::{Client, Response};
-
-/// Should only be implemented by `List*` parameter types. Kept public so that
-/// the generated code crates can access this trait.
-#[doc(hidden)]
-pub trait PaginationParams: Serialize {}
+use crate::Client;
 
 #[derive(Debug)]
 pub struct ListPaginator<T> {
-    data: Vec<T>,
-    url: String,
-    has_more: bool,
-    total_count: Option<u64>,
+    page: T,
     params: serde_json::Value,
 }
 
-pub trait PaginationExt<T> {
-    fn into_paginator(self) -> ListPaginator<T>;
+pub trait PaginationExt {
+    type Data;
+
+    fn into_paginator(self) -> ListPaginator<Self::Data>;
 }
 
-impl<T> PaginationExt<T> for List<T>
+impl<T> PaginationExt for List<T>
 where
-    T: Object + DeserializeOwned + Send + Sync + 'static,
-    T::Id: ToString,
+    T: Sync + Send + 'static,
+    List<T>: PaginableList,
 {
-    fn into_paginator(self) -> ListPaginator<T> {
-        let mut paginator = ListPaginator {
-            data: self.data,
-            // the url we get back is prefixed
-            url: self.url.trim_start_matches("/v1/").to_string(),
-            has_more: self.has_more,
-            total_count: self.total_count,
-            params: Default::default(),
-        };
-        if let Some(curr_cursor) = paginator.data.last().and_then(|t| t.id().as_cursor_opt()) {
-            paginator.update_cursor(curr_cursor.to_string());
-        }
-        paginator
+    type Data = List<T>;
+
+    fn into_paginator(mut self) -> ListPaginator<List<T>> {
+        let mut params = Default::default();
+        self.update_params(&mut params);
+        ListPaginator { page: self, params }
     }
 }
 
-impl<T> ListPaginator<T> {
-    /// Kept public so that the generated code crates can access this trait. Used by `List*` params
-    /// to implement `PaginationExt`.
+impl<T> PaginationExt for SearchList<T>
+where
+    T: Sync + Send + 'static,
+    SearchList<T>: PaginableList,
+{
+    type Data = SearchList<T>;
+
+    fn into_paginator(mut self) -> ListPaginator<SearchList<T>> {
+        let mut params = Default::default();
+        self.update_params(&mut params);
+        ListPaginator { page: self, params }
+    }
+}
+
+impl<T> ListPaginator<SearchList<T>> {
+    /// Kept public so that the generated code crates can access this trait. Used by `Search*` params
+    /// to implement `PaginationExt`. Not part of the public API.
     #[doc(hidden)]
-    pub fn from_params(url: &str, params: impl PaginationParams) -> Self {
-        ListPaginator {
-            data: vec![],
+    pub fn from_search_params(url: &str, params: impl Serialize) -> Self {
+        let page = SearchList {
             url: url.to_string(),
             has_more: true,
+            data: vec![],
+            next_page: None,
             total_count: None,
-            params: serde_json::to_value(params).expect("Invalid pagination params"),
+        };
+        Self {
+            page,
+            params: serde_json::to_value(params)
+                // Expect should be safe since we control which types call this
+                .expect("all Stripe types implement `Serialize` infallibly"),
+        }
+    }
+}
+
+impl<T> ListPaginator<List<T>> {
+    /// Kept public so that the generated code crates can access this trait. Used by `List*` params
+    /// to implement `PaginationExt`. Not part of the public API.
+    #[doc(hidden)]
+    pub fn from_list_params(url: &str, params: impl Serialize) -> Self {
+        let page = List { data: vec![], has_more: true, total_count: None, url: url.to_string() };
+        Self {
+            page,
+            params: serde_json::to_value(params)
+                .expect("all Stripe types implement `Serialize` infallibly"),
         }
     }
 }
 
 impl<T> ListPaginator<T>
 where
-    T: Object + DeserializeOwned + Send + Sync + 'static,
+    T: PaginableList + Send + Sync + 'static,
 {
     /// Repeatedly queries Stripe for more data until all elements in list are fetched, using
     /// Stripe's default page size.
     ///
     /// Requires `feature = "blocking"`.
     #[cfg(feature = "blocking")]
-    pub fn get_all(self, client: &Client) -> Response<Vec<T>> {
-        let mut data = Vec::with_capacity(self.total_count.unwrap_or(0) as usize);
-        let mut paginator = self;
+    pub fn get_all(self, client: &Client) -> crate::Response<Vec<T::Data>> {
+        let mut data = vec![];
+        let mut parts = self.page.into_parts();
+        let mut params = self.params;
         loop {
-            if !paginator.has_more {
-                data.extend(paginator.data);
+            // `append` empties `parts.data`
+            data.append(&mut parts.data);
+
+            if !parts.has_more {
                 break;
             }
-            let next_page = paginator.fetch_page_with_curr_params(client)?;
-            paginator.update_with_new_data(next_page);
+
+            let mut next_page: T =
+                client.get_query(parts.url.trim_start_matches("/v1/"), &params)?;
+            next_page.update_params(&mut params);
+            parts = next_page.into_parts();
         }
         Ok(data)
     }
@@ -92,67 +118,59 @@ where
     /// Requires `feature = ["async", "stream"]`.
     #[cfg(all(feature = "async", feature = "stream"))]
     pub fn stream(
-        mut self,
+        self,
         client: &Client,
-    ) -> impl futures_util::Stream<Item = Result<T, crate::StripeError>> + Unpin {
+    ) -> impl futures_util::Stream<Item = Result<T::Data, crate::StripeError>> + Unpin {
         // We are going to be popping items off the end of the list, so we need to reverse it.
-        self.data.reverse();
+        let mut page = self.page.into_parts();
+        page.data.reverse();
+        let paginator = ListPaginator { page: T::from_parts(page), params: self.params };
 
-        Box::pin(futures_util::stream::unfold(Some((self, client.clone())), Self::unfold_stream))
+        Box::pin(futures_util::stream::unfold(
+            Some((paginator, client.clone())),
+            Self::unfold_stream,
+        ))
     }
 
     /// unfold a single item from the stream
     #[cfg(all(feature = "async", feature = "stream"))]
     async fn unfold_stream(
         state: Option<(Self, Client)>,
-    ) -> Option<(Result<T, crate::StripeError>, Option<(Self, Client)>)> {
-        let (mut paginator, client) = state?; // If none, we sent the last item in the last iteration
-
-        if let Some(next_val) = paginator.data.pop() {
+    ) -> Option<(Result<T::Data, crate::StripeError>, Option<(Self, Client)>)> {
+        let (paginator, client) = state?; // If none, our last request was an error, so stop here
+        let mut parts = paginator.page.into_parts();
+        if let Some(next_val) = parts.data.pop() {
             // We have more data on this page
-            return Some((Ok(next_val), Some((paginator, client))));
+            return Some((
+                Ok(next_val),
+                Some((Self { page: T::from_parts(parts), params: paginator.params }, client)),
+            ));
         }
 
         // Final value of the stream, no errors
-        if !paginator.has_more {
+        if !parts.has_more {
             return None;
         }
 
-        match paginator.fetch_page_with_curr_params(&client).await {
-            Ok(next_page) => {
-                debug_assert!(paginator.data.is_empty());
-                paginator.update_with_new_data(next_page);
+        match client
+            .get_query::<T, _>(parts.url.trim_start_matches("/v1/"), &paginator.params)
+            .await
+        {
+            Ok(mut next_page) => {
+                let mut params = paginator.params;
+                next_page.update_params(&mut params);
+
+                let mut parts = next_page.into_parts();
 
                 // We are going to be popping items off the end of the list, so we need to reverse it.
-                // The assert above ensures we are only reversing this specific list we've
-                // just received
-                paginator.data.reverse();
+                parts.data.reverse();
 
-                let next_val = paginator.data.pop()?;
+                let next_val = parts.data.pop()?;
 
                 // Yield last value of this page, the next page (and client) becomes the state
-                Some((Ok(next_val), Some((paginator, client))))
+                Some((Ok(next_val), Some((Self { page: T::from_parts(parts), params }, client))))
             }
-            Err(e) => Some((Err(e), None)), // We ran into an error. The last value of the stream will be the error.
+            Err(err) => Some((Err(err), None)), // We ran into an error. The last value of the stream will be the error.
         }
-    }
-
-    fn fetch_page_with_curr_params(&self, client: &Client) -> Response<List<T>> {
-        client.get_query(&self.url, &self.params)
-    }
-
-    fn update_cursor(&mut self, id: String) {
-        self.params["starting_after"] = serde_json::Value::String(id);
-    }
-
-    fn update_with_new_data(&mut self, list: List<T>) {
-        self.has_more = list.has_more;
-        self.total_count = list.total_count;
-        if let Some(new_cursor) = list.data.last().and_then(|l| l.id().as_cursor_opt()) {
-            self.update_cursor(new_cursor.to_string());
-        } else {
-            self.has_more = false;
-        }
-        self.data.extend(list.data);
     }
 }