+++ tests/tests/it/main.rs
@@ -8,3 +8,4 @@ mod price;
 mod async_tests;
 #[cfg(feature = "blocking")]
 mod blocking;
+mod pagination_utils;